Rename folders, update Readme

+++ lab-3/kotiki-java/service/src/main/java/ru/itmo/kotiki/service/controller/CatController.java
@@ -1,11 +1,19 @@
 package ru.itmo.kotiki.service.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import ru.itmo.kotiki.dao.entity.CatColor;
 import ru.itmo.kotiki.service.dto.CatDto;
 import ru.itmo.kotiki.service.service.CatService;
+import ru.itmo.kotiki.service.service.UserService;
+import ru.itmo.kotiki.service.userDetails.UserRole;
+
+import java.util.Collection;
 
 @RestController
 @RequestMapping("/cats/")
@@ -14,67 +22,105 @@ public class CatController {
     @Autowired
     private CatService catService;
 
+    @Autowired
+    private UserService userService;
+
     @GetMapping("all")
-    public ResponseEntity AllCats() {
+    public ResponseEntity<Object> AllCats() {
         try {
-            return ResponseEntity.ok(catService.getAllCats());
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                return ResponseEntity.ok(catService.getAllCats());
+            } else {
+                return ResponseEntity.ok(userService.getAllCats());
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
 
     @GetMapping("id")
-    public ResponseEntity ById(@RequestParam String id) {
+    public ResponseEntity<Object> ById(@RequestParam String id) {
         try {
-            return ResponseEntity.ok(catService.findById(id));
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                return ResponseEntity.ok(catService.findById(id));
+            } else {
+                return new ResponseEntity<>("", HttpStatus.FORBIDDEN);
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
 
     @GetMapping("id/owner")
-    public ResponseEntity OwnerByCatId(@RequestParam String id) {
+    public ResponseEntity<Object> OwnerByCatId(@RequestParam String id) {
         try {
-            return ResponseEntity.ok(catService.findOwnerByCatId(id));
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                return ResponseEntity.ok(catService.findOwnerByCatId(id));
+            } else {
+                return new ResponseEntity<>("", HttpStatus.FORBIDDEN);
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
 
     @GetMapping("color")
-    public ResponseEntity CatsWithCatColor(@RequestParam CatColor color) {
+    public ResponseEntity<Object> CatsWithCatColor(@RequestParam CatColor color) {
         try {
-            return ResponseEntity.ok(catService.getCatsWithCatColor(color));
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                return ResponseEntity.ok(catService.getCatsWithCatColor(color));
+            } else {
+                return ResponseEntity.ok(userService.getCatsWithCatColor(color));
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
 
     @PostMapping("save")
-    public ResponseEntity save(@RequestBody CatDto catDto) {
+    public ResponseEntity<Object> save(@RequestBody CatDto catDto) {
         try {
-            catService.save(catDto);
-            return ResponseEntity.ok("Cat successfully saved.");
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                catService.save(catDto);
+                return ResponseEntity.ok("Cat successfully saved.");
+            } else {
+                return new ResponseEntity<>("", HttpStatus.FORBIDDEN);
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
 
     @PostMapping("id/owner")
-    public ResponseEntity setOwnerById(@RequestParam String catId, @RequestParam String ownerId) {
+    public ResponseEntity<Object> setOwnerById(@RequestParam String catId, @RequestParam String ownerId) {
         try {
-            catService.setOwnerById(catId, ownerId);
-            return ResponseEntity.ok("Successfully set new owner.");
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                catService.setOwnerById(catId, ownerId);
+                return ResponseEntity.ok("Successfully set new owner.");
+            } else {
+                return new ResponseEntity<>("", HttpStatus.FORBIDDEN);
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
 
     @DeleteMapping("id")
-    public ResponseEntity delete(@RequestParam String id) {
+    public ResponseEntity<Object> delete(@RequestParam String id) {
         try {
-            catService.delete(id);
-            return ResponseEntity.ok("Cat successfully deleted.");
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                catService.delete(id);
+                return ResponseEntity.ok("Cat successfully deleted.");
+            } else {
+                return new ResponseEntity<>("", HttpStatus.FORBIDDEN);
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
