+++ lab-3/kotiki-java/service/src/main/java/ru/itmo/kotiki/service/controller/OwnerController.java
@@ -1,10 +1,17 @@
 package ru.itmo.kotiki.service.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import ru.itmo.kotiki.service.dto.OwnerDto;
 import ru.itmo.kotiki.service.service.OwnerService;
+import ru.itmo.kotiki.service.userDetails.UserRole;
+
+import java.util.Collection;
 
 @RestController
 @RequestMapping("/owners/")
@@ -16,7 +23,12 @@ public class OwnerController {
     @GetMapping("all")
     public ResponseEntity AllOwners() {
         try {
-            return ResponseEntity.ok(ownerService.getAllOwners());
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                return ResponseEntity.ok(ownerService.getAllOwners());
+            } else {
+                return new ResponseEntity<>("", HttpStatus.FORBIDDEN);
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
@@ -25,7 +37,12 @@ public class OwnerController {
     @GetMapping("cats")
     public ResponseEntity CatsByOwnerId(@RequestParam String ownerId) {
         try {
-            return ResponseEntity.ok(ownerService.getCatsByOwnerId(ownerId));
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                return ResponseEntity.ok(ownerService.getCatsByOwnerId(ownerId));
+            } else {
+                return new ResponseEntity<>("", HttpStatus.FORBIDDEN);
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
@@ -34,8 +51,13 @@ public class OwnerController {
     @PostMapping("save")
     public ResponseEntity save(@RequestBody OwnerDto ownerDto) {
         try {
-            ownerService.save(ownerDto);
-            return ResponseEntity.ok("Owner successfully saved.");
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                ownerService.save(ownerDto);
+                return ResponseEntity.ok("Owner successfully saved.");
+            } else {
+                return new ResponseEntity<>("", HttpStatus.FORBIDDEN);
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
@@ -44,7 +66,12 @@ public class OwnerController {
     @GetMapping("id")
     public ResponseEntity ById(@RequestParam String id) {
         try {
-            return ResponseEntity.ok(ownerService.findById(id));
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                return ResponseEntity.ok(ownerService.findById(id));
+            } else {
+                return new ResponseEntity<>("", HttpStatus.FORBIDDEN);
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
@@ -53,8 +80,13 @@ public class OwnerController {
     @DeleteMapping("id")
     public ResponseEntity delete(@RequestParam String id) {
         try {
-            ownerService.delete(id);
-            return ResponseEntity.ok("Owner successfully deleted.");
+            Collection<? extends GrantedAuthority> authorities = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+            if (authorities.contains(new SimpleGrantedAuthority(UserRole.ADMIN.toString()))){
+                ownerService.delete(id);
+                return ResponseEntity.ok("Owner successfully deleted.");
+            } else {
+                return new ResponseEntity<>("", HttpStatus.FORBIDDEN);
+            }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
